+++ tests/siteadapter_test.tcl
@@ -51,6 +51,16 @@ puts "Parsed nxPre PRE-FLAC: [dict get $parsedFlac payload]"
 
 set newDirLine {07-02-2026 10:00:32 NEWDIR: "khaz" "MEV" "/GROUPS/MEV/MUSiC/DEATH_IN_AUGUST-GREET_THE_STORM-EP-CD-FLAC-2026-TOTENKVLT" "\\?\C:\ioFTPD\FTP-ROOT-DIR\GROUPS\MEV\MUSiC\DEATH_IN_AUGUST-GREET_THE_STORM-EP-CD-FLAC-2026-TOTENKVLT"}
 set tvNewDirLine {07-02-2026 10:01:32 NEWDIR: "khaz" "MEV" "/GROUPS/MEV/TV/The.Last.of.Us.S01E01.1080p.WEB.H264-GROUP" "\\?\C:\ioFTPD\FTP-ROOT-DIR\GROUPS\MEV\TV\The.Last.of.Us.S01E01.1080p.WEB.H264-GROUP"}
+set firstLine {07-02-2026 10:01:35 UPDATE_MOVIE: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Vroom GROUP 81 10680 23708262 file001.rar Twisters.2024.1080p.BluRay.H264-VALUE tagline {9m 9s}}
+set halfLine {07-02-2026 10:02:35 HALFWAY_RACE_MOVIE: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Twisters.2024.1080p.BluRay.H264-VALUE Vroom GROUP 8806 23 79 8376 GROUP 8806 23 79 8376 23708262 81 79 8376 4010 17 file001.rar 2 1 Transporter GROUP 81 {{Transporter GROUP} {ioftpd NoGroup}} {9m 9s}}
+set racerLine {07-02-2026 10:03:35 RACE_MOVIE: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Transporter GROUP {{Vroom GROUP}} Twisters.2024.1080p.BluRay.H264-VALUE 4096 file002.rar 12 42 1 1 34 81 47 tagline {{Vroom GROUP}} {8m 20s}}
+set leaderLine {07-02-2026 10:04:35 NEWLEADER_MOVIE: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Transporter GROUP 4096 12 34 42 23708262 Twisters.2024.1080p.BluRay.H264-VALUE file002.rar 1 1 81 47 Transporter GROUP 1000 34 42 4096 GROUP 1000 34 42 4096 tagline {{Vroom GROUP}} {8m 20s}}
+set badCrcLine {07-02-2026 10:05:35 BAD_FILE_CRC: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Twisters.2024.1080p.BluRay.H264-VALUE Transporter GROUP broken.r00}
+set zeroLine {07-02-2026 10:06:35 BAD_FILE_0SIZE: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Twisters.2024.1080p.BluRay.H264-VALUE Transporter GROUP empty.r00}
+set nfoLine {07-02-2026 10:07:35 NFO: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Transporter GROUP Twisters.2024.1080p.BluRay.H264-VALUE twisters.nfo}
+set doubleSfvLine {07-02-2026 10:08:35 DOUBLESFV: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Transporter GROUP Twisters.2024.1080p.BluRay.H264-VALUE twisters.sfv}
+set speedTestLine {07-02-2026 10:09:35 SPEEDTEST: {/SPEEDTEST/} Transporter GROUP tagline 20480 512}
+set incompleteLine {07-02-2026 10:10:35 INCOMPLETE: {/MOVIE-1080P/Twisters.2024.1080p.BluRay.H264-VALUE/} Transporter GROUP Twisters.2024.1080p.BluRay.H264-VALUE}
 set parsedNewDir [::dZSbot::SiteAdapter::ParseIoFtpdLine $newDirLine]
 if {![dict get $parsedNewDir ok] || [dict get $parsedNewDir event] ne "site.newdir"} {
     error "Expected ioFTPD NEWDIR line to parse"
@@ -69,6 +79,50 @@ if {[dict get $parsedTvNewDir payload section] ne "TV"} {
 }
 puts "Parsed ioFTPD TV NEWDIR: [dict get $parsedTvNewDir payload]"
 
+set parsedFirst [::dZSbot::SiteAdapter::ParseIoFtpdLine $firstLine]
+if {![dict get $parsedFirst ok] || [dict get $parsedFirst event] ne "site.upload.first"} {
+    error "Expected ioFTPD UPDATE line to parse as FIRST"
+}
+set firstAnnounce [::dZSbot::Modules::Upload::FormatLine first [dict get $parsedFirst payload]]
+if {[string first "Twisters.2024.1080p.BluRay.H264-VALUE by Vroom - 10.43MB/s :: release size - 22.61GB ::" $firstAnnounce] < 0} {
+    error "Expected FIRST announce format: $firstAnnounce"
+}
+puts "Parsed ioFTPD FIRST: [dict get $parsedFirst payload]"
+
+set parsedHalf [::dZSbot::SiteAdapter::ParseIoFtpdLine $halfLine]
+if {![dict get $parsedHalf ok] || [dict get $parsedHalf event] ne "site.upload.half"} {
+    error "Expected ioFTPD HALFWAY line to parse as HALF"
+}
+set halfAnnounce [::dZSbot::Modules::Upload::FormatLine half [dict get $parsedHalf payload]]
+if {[string first "Twisters.2024.1080p.BluRay.H264-VALUE :: Vroom :: with 23f :: 79% :: 8.60GB :: 8.18MB/s :: others: Transporter/GROUP, ioftpd/NoGroup :: time left 9m 9s" $halfAnnounce] < 0} {
+    error "Expected HALF announce format: $halfAnnounce"
+}
+puts "Parsed ioFTPD HALF: [dict get $parsedHalf payload]"
+
+foreach item [list \
+    [list $racerLine site.upload.racer "Twisters.2024.1080p.BluRay.H264-VALUE :: Transporter :: 4.00MB/s"] \
+    [list $leaderLine site.upload.leader "Twisters.2024.1080p.BluRay.H264-VALUE :: Transporter"] \
+    [list $badCrcLine site.upload.badfile "badcrc" "broken.r00"] \
+    [list $zeroLine site.upload.badfile "0size" "empty.r00"] \
+    [list $nfoLine site.upload.nfo "NFO" "twisters.nfo"] \
+    [list $doubleSfvLine site.upload.doublesfv "DOUBLESFV" "twisters.sfv"] \
+    [list $speedTestLine site.upload.speedtest "512MB :: 20.00MB/s"] \
+    [list $incompleteLine site.upload.incomplete "INCOMPLETE" "Twisters.2024.1080p.BluRay.H264-VALUE"]] {
+    set parsedExtra [::dZSbot::SiteAdapter::ParseIoFtpdLine [lindex $item 0]]
+    set expectedEvent [lindex $item 1]
+    if {![dict get $parsedExtra ok] || [dict get $parsedExtra event] ne $expectedEvent} {
+        error "Expected $expectedEvent from extra ioFTPD line: $parsedExtra"
+    }
+    set action [dict get [dict get $parsedExtra payload] action]
+    set line [::dZSbot::Modules::Upload::FormatLine $action [dict get $parsedExtra payload]]
+    foreach expected [lrange $item 2 end] {
+        if {[string first $expected $line] < 0} {
+            error "Expected '$expected' in $expectedEvent announce: $line"
+        }
+    }
+    puts "Parsed extra ioFTPD event $expectedEvent: [dict get $parsedExtra payload]"
+}
+
 set completeLine {07-02-2026 10:00:43 COMPLETE_STAT_RACE_FLAC: /GROUPS/MEV/MUSiC/DEATH_IN_AUGUST-GREET_THE_STORM-EP-CD-FLAC-2026-TOTENKVLT/ DEATH_IN_AUGUST-GREET_THE_STORM-EP-CD-FLAC-2026-TOTENKVLT 82865 5 7533 234909 11 1 1 khaz NoGroup 227156}
 set parsedComplete [::dZSbot::SiteAdapter::ParseIoFtpdLine $completeLine]
 if {![dict get $parsedComplete ok] || [dict get $parsedComplete event] ne "site.upload.complete"} {
@@ -80,7 +134,21 @@ if {[dict get $parsedComplete payload files] ne "5"} {
 if {[dict get $parsedComplete payload user] ne "khaz" || [dict get $parsedComplete payload group] ne "NoGroup"} {
     error "Expected COMPLETE uploader khaz/NoGroup"
 }
+if {[dict get $parsedComplete payload speed_kbps] ne "7533"} {
+    error "Expected COMPLETE average speed"
+}
+set completeAnnounce [::dZSbot::Modules::Upload::FormatLine complete [dict get $parsedComplete payload]]
+if {[string first "DEATH_IN_AUGUST-GREET_THE_STORM-EP-CD-FLAC-2026-TOTENKVLT :: 5f :: 80.92MB :: 11s :: avg. 7.36MB/s :: khaz" $completeAnnounce] < 0} {
+    error "Expected COMPLETE announce to include upload speed: $completeAnnounce"
+}
 puts "Parsed ioFTPD COMPLETE: [dict get $parsedComplete payload]"
+
+if {![::dZSbot::Modules::Upload::SectionAllowed TV-1080P]} {
+    error "Expected upload announce to allow TV-1080P"
+}
+if {![::dZSbot::Modules::Upload::SectionAllowed MOVIE-2160P]} {
+    error "Expected upload announce to allow MOVIE-2160P"
+}
 
 set preTestFile [file join $root runtime test-nxpre.tsv]
 catch {file delete $preTestFile}